Add word censoring and blocklisting
Include separate word list for words to censor and words to avoid using
in metadata like titles and tags. This is done as there is not necessarily
a 1:1 overlap with words to censor within videos and words that will cause
search engines and video hosting sites to de-prioritize a video for reasons
such as advertising and monetization.