Initialize word filters outside of suites
Word filters/censors should eventually be wrapped into a custom class.
Currently, it is possible that filters will be initialized without
actually being used. A wrapper class could wait until the filter is first
used to initialize the internal filter. This is only a performance concern
with very large lists of words for filters.